refactor: move src/features/observe/accessibility/ to android/
Rename the accessibility directory to android to better reflect that it
contains Android-specific observation functionality rather than being
named after the implementation detail (Accessibility Service).

Fixes #1123
