+++ src/features/observe/AccessibilityServiceClient.ts
@@ -2,16 +2,16 @@
  * AccessibilityServiceClient - Backward compatibility re-export
  *
  * This file re-exports from the new modular location for backward compatibility.
- * New code should import from 'src/features/observe/accessibility' directly.
+ * New code should import from 'src/features/observe/android' directly.
  *
- * @deprecated Import from 'src/features/observe/accessibility' instead
+ * @deprecated Import from 'src/features/observe/android' instead
  */
 
 // Re-export main client
-export { AccessibilityServiceClient } from "./accessibility";
+export { AccessibilityServiceClient } from "./android";
 
 // Re-export interfaces
-export type { AccessibilityService, InteractionEvent } from "./accessibility";
+export type { AccessibilityService, InteractionEvent } from "./android";
 
 // Re-export result types
 export type {
@@ -31,4 +31,4 @@ export type {
   A11yDeviceOwnerStatusResult,
   A11yPermissionResult,
   AndroidPerfTiming,
-} from "./accessibility";
+} from "./android";
