+++ src/features/observe/android/AccessibilityServiceGestures.ts
